+++ src/libs/mod.rs
@@ -1,5 +1,4 @@
 pub mod colored;
 pub mod os_type;
 pub mod path;
-pub mod prompt;
 pub mod re;