+++ src/builtins/export.rs
@@ -3,10 +3,10 @@ use std::env;
 use std::io::Write;
 use types::Tokens;
 
+use libs;
 use parsers;
 use shell;
 use tools;
-use libs;
 
 pub fn run(_sh: &shell::Shell, tokens: &Tokens) -> i32 {
     let mut i = 0;