+++ src/builtins/mod.rs
@@ -1,9 +1,9 @@
+pub mod bg;
 pub mod cd;
 pub mod cinfo;
 pub mod exec;
 pub mod exit;
 pub mod export;
-pub mod bg;
 pub mod fg;
 pub mod history;
 pub mod jobs;