+++ src/prompt/mod.rs
@@ -6,7 +6,6 @@ use shell;
 use self::main::get_prompt_string;
 use self::main::render_prompt;
 
-
 pub fn get_prompt(sh: &shell::Shell) -> String {
     let ps = get_prompt_string();
     render_prompt(sh, &ps)