+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Change Logs
 
+## 0.9.1
+
+- Skip hidden files when expanding `foo/*`.
+
 ## v0.9.0
 
 - Works on escape file names.
 - Some improvements on command line parser (escape chars etc).
 - Added suport for customizing prompt.
 
-## v0.8.9 - 2018-11-11
+## v0.8.9
 
 - Fixed issue of finding command in `$PATH`.
 - Fixed issue of cmds like `(ls)`.